Разделил код на модули.
