+++ js/function.js
@@ -43,3 +43,24 @@ exctractNumber('ECMAScript 2022');
 exctractNumber('1 кефир, 0.5 батона');
 exctractNumber('агент 007');
 exctractNumber('а я томат');
+
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) =>
+  elements[getRandomInteger(0, elements.length - 1)];
+
+function createIdGenerator() {
+  let lastGeneratedId = 0;
+
+  return () => {
+    lastGeneratedId++;
+    return lastGeneratedId;
+  };
+}
+
+export { getRandomInteger, getRandomArrayElement, createIdGenerator };
